+++ docs/fr_FR/changelog.md
@@ -6,6 +6,10 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
+# 29/07/2026
+
+- Revue de la page de configuration du plugin
+
 # 29/06/2026
 
 - Ajout du support de UCG Max